+++ modules/engine/src/index.ts
@@ -121,8 +121,6 @@ export {DynamicTexture} from './dynamic-texture/dynamic-texture';
 export type {TextureBindingSource} from './dynamic-texture/texture-binding-source';
 export type {VideoTextureProps, VideoTextureSource} from './dynamic-texture/video-texture';
 export {VideoTexture} from './dynamic-texture/video-texture';
-export type {HTMLTextureProps} from './dynamic-texture/html-texture';
-export {HTMLTexture} from './dynamic-texture/html-texture';
 export type {
   BufferBindingSource,
   DynamicBufferProps,