+++ modules/experimental/src/index.ts
@@ -5,3 +5,6 @@
 export {type TextureFormatPacked, RGBADecoder} from './textures/rgba-decoder';
 
 export {TEXTURE_FORMAT_PIXEL_DECODERS} from './textures/packed-pixels';
+
+export type {HTMLTextureProps} from './textures/html-texture';
+export {HTMLTexture} from './textures/html-texture';